+++ src/codecs/theora.cpp
@@ -68,11 +68,11 @@ void TheoraStream::readNextFrame(YCbCrBuffer &ycbcrBuffer) {
 	th_ycbcr_buffer buffer;
 	th_decode_ycbcr_out(_decoder, buffer);
 
-	if (ycbcrBuffer.y.size() != buffer[0].width * buffer[0].height)
+	if (ycbcrBuffer.y.size() != static_cast<unsigned int>(buffer[0].width * buffer[0].height))
 		ycbcrBuffer.y.resize(buffer[0].width * buffer[0].height);
-	if (ycbcrBuffer.cb.size() != buffer[1].width * buffer[1].height)
+	if (ycbcrBuffer.cb.size() != static_cast<unsigned int>(buffer[1].width * buffer[1].height))
 		ycbcrBuffer.cb.resize(buffer[1].width * buffer[1].height);
-	if (ycbcrBuffer.cr.size() != buffer[2].width * buffer[2].height)
+	if (ycbcrBuffer.cr.size() != static_cast<unsigned int>(buffer[2].width * buffer[2].height))
 		ycbcrBuffer.cr.resize(buffer[2].width * buffer[2].height);
 
 	for (int i = 0; i < buffer[0].height; ++i) {
